+++ src/main/java/SearchResult.java
@@ -1,9 +1,9 @@
 public class SearchResult {
 
     private final CodeChunk chunk;
-    private final int score;
+    private final double score;
 
-    public SearchResult(CodeChunk chunk, int score) {
+    public SearchResult(CodeChunk chunk, double score) {
         this.chunk = chunk;
         this.score = score;
     }
@@ -12,7 +12,7 @@ public class SearchResult {
         return chunk;
     }
 
-    public int getScore() {
+    public double getScore() {
         return score;
     }
 }
